JAVA-29238 | fixing tests (#15434)
