+++ persistence-modules/spring-data-jpa-repo/README.md
@@ -20,6 +20,6 @@ After importing the project into Eclipse, you may see the following error:
 "No persistence xml file found in project"
 
 This can be ignored: 
-- Project -> Properties -> Java Persistance -> JPA -> Error/Warnings -> Select Ignore on "No persistence xml file found in project"
+- Project -> Properties -> Java Persistence -> JPA -> Error/Warnings -> Select Ignore on "No persistence xml file found in project"
 Or: 
 - Eclipse -> Preferences - Validation - disable the "Build" execution of the JPA Validator
