+++ persistence-modules/spring-data-jpa-repo/src/main/java/com/baeldung/jpa/config/PersistenceConfig.java
@@ -1,4 +1,4 @@
-package com.baeldung.repository;
+package com.baeldung.jpa.config;
 
 import java.util.Properties;
 
@@ -6,12 +6,10 @@ import javax.sql.DataSource;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.core.env.Environment;
 import org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
@@ -23,24 +21,18 @@ import com.google.common.base.Preconditions;
 
 @Configuration
 @PropertySource("classpath:persistence.properties")
-@ComponentScan("com.baeldung.repository")
-//@ImportResource("classpath*:*springDataConfig.xml")
 @EnableTransactionManagement
-@EnableJpaRepositories(basePackages = "com.baeldung.repository")
+//@ImportResource("classpath*:*springDataConfig.xml")
 public class PersistenceConfig {
 
     @Autowired
     private Environment env;
 
-    public PersistenceConfig() {
-        super();
-    }
-
     @Bean
     public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
         final LocalContainerEntityManagerFactoryBean em = new LocalContainerEntityManagerFactoryBean();
         em.setDataSource(dataSource());
-        em.setPackagesToScan("com.baeldung.spring.data.persistence.repository");
+        em.setPackagesToScan("com.baeldung.jpa.domain");
 
         final HibernateJpaVendorAdapter vendorAdapter = new HibernateJpaVendorAdapter();
         em.setJpaVendorAdapter(vendorAdapter);
