+++ persistence-modules/spring-data-jpa-repo/src/main/java/com/baeldung/jpa/domain/Song.java
@@ -1,4 +1,4 @@
-package com.baeldung.entity;
+package com.baeldung.jpa.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
