+++ persistence-modules/spring-data-jpa-repo/src/main/java/com/baeldung/jpa/repository/FruitRepository.java
@@ -1,0 +1,27 @@
+package com.baeldung.jpa.repository;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+import com.baeldung.jpa.domain.Fruit;
+
+@Repository
+public interface FruitRepository extends JpaRepository<Fruit, Long> {
+
+    Long deleteByName(String name);
+
+    List<Fruit> deleteByColor(String color);
+
+    Long removeByName(String name);
+
+    List<Fruit> removeByColor(String color);
+
+    @Modifying
+    @Query("delete from Fruit f where f.name=:name or f.color=:color")
+    int deleteFruits(@Param("name") String name, @Param("color") String color);
+}
