+++ persistence-modules/spring-data-jpa-repo/src/main/java/com/baeldung/jpa/repository/IFooDAO.java
@@ -1,9 +1,10 @@
-package com.baeldung.repository;
+package com.baeldung.jpa.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
+
+import com.baeldung.jpa.domain.Foo;
 
 public interface IFooDAO extends JpaRepository<Foo, Long> {
