+++ persistence-modules/spring-data-jpa-repo/src/main/java/com/baeldung/jpa/repository/PassengerRepository.java
@@ -1,13 +1,13 @@
-package com.baeldung.repository;
+package com.baeldung.jpa.repository;
 
-import com.baeldung.entity.Passenger;
+import com.baeldung.jpa.domain.Passenger;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-interface PassengerRepository extends JpaRepository<Passenger, Long> {
+public interface PassengerRepository extends JpaRepository<Passenger, Long> {
 
     List<Passenger> findByFirstNameIgnoreCase(String firstName);
