+++ persistence-modules/spring-data-jpa-repo/src/main/java/com/baeldung/jpa/service/FooService.java
@@ -1,10 +1,14 @@
-package com.baeldung.repository;
+package com.baeldung.jpa.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.baeldung.jpa.domain.Foo;
+import com.baeldung.jpa.repository.IFooDAO;
+
 @Service
 public class FooService implements IFooService {
+
     @Autowired
     private IFooDAO dao;
