+++ persistence-modules/spring-data-jpa-repo/src/main/java/com/baeldung/jpa/service/IFooService.java
@@ -1,0 +1,7 @@
+package com.baeldung.jpa.service;
+
+import com.baeldung.jpa.domain.Foo;
+
+public interface IFooService {
+    Foo create(Foo foo);
+}
