--- persistence-modules/spring-data-jpa-repo/src/main/java/com/baeldung/repository/IFooService.java
@@ -1,5 +1,0 @@
-package com.baeldung.repository;
-
-public interface IFooService {
-    Foo create(Foo foo);
-}
