+++ persistence-modules/spring-data-jpa-repo/src/main/resources/fruit-data.json
@@ -1,12 +1,12 @@
 [
     {
-        "_class": "com.baeldung.entity.Fruit",
+        "_class": "com.baeldung.jpa.domain.Fruit",
         "name": "apple",
         "color": "red",
         "id": 1
     },
     {
-        "_class": "com.baeldung.entity.Fruit",
+        "_class": "com.baeldung.jpa.domain.Fruit",
         "name": "guava",
         "color": "green",
         "id": 2
